+++ src/main/java/com/invadermonky/magicultureintegrations/MagicultureIntegrations.java
@@ -19,7 +19,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class MagicultureIntegrations {
     public static final String MOD_ID = "magicultureintegrations";
     public static final String MOD_NAME = "Magiculture Integrations";
-    public static final String MOD_VERSION = "1.12.2-2.1.1";
+    public static final String MOD_VERSION = "1.12.2-2.1.2";
     public static final String MC_VERSION = "[1.12.2]";
     public static final String DEPENDENCIES =
             ";required-after:mixinbooter@[9.0,)" +